Completed xlsio functionality and docs
- Added functionality to indicate reverse biased cells. Reverse biased and bypassed cells are indicated on the same excel tab (assuming that a bypassed cell cannot be reverse biased).
- Added functionality of temperature input.
- Fixed imports and style.
- Updated README, example_workflow .py and .ipynb.
