+++ pvmismatch/contrib/xlsio/__init__.py
@@ -1,0 +1,2 @@
+# API for xlsio
+from pvmismatch.contrib.xlsio.xlsio import *
